refac: Combine into single question.go
