--- internal/tui/question/init.go
@@ -1,7 +1,0 @@
-package question
-
-import tea "github.com/charmbracelet/bubbletea"
-
-func (qm QuestionModel) Init() tea.Cmd {
-	return nil
-}
