--- internal/tui/question/view.go
@@ -1,5 +1,0 @@
-package question
-
-func (qm QuestionModel) View() string {
-	return "\n" + m.list.View()
-}
